+++ src/lse/events/BlockEvents.cpp
@@ -1,5 +1,6 @@
 #include "legacy/api/BaseAPI.h"
 #include "legacy/api/BlockAPI.h"
+#include "legacy/api/ContainerAPI.h"
 #include "legacy/api/EntityAPI.h"
 #include "legacy/api/EventAPI.h"
 #include "legacy/api/ItemAPI.h"
@@ -518,6 +519,40 @@ LL_TYPE_INSTANCE_HOOK(
     return origin(region, commandOrigin, markForSaving);
 }
 
+namespace dispenser {
+LL_TYPE_INSTANCE_HOOK(
+    DispenserEjectItemHook,
+    HookPriority::Normal,
+    DispenserBlock,
+   &DispenserBlock::ejectItem,
+    void,
+    BlockSource&     region,
+    Vec3 const&      pos,
+    uchar            face,
+    ItemStack const& item,
+    Container&       container,
+    int              slot,
+    int              countLimit
+) {
+    IF_LISTENED(EVENT_TYPES::onDispenseItem) {
+        if (checkClientIsServerThread()) {
+            if (!CallEvent(
+                    EVENT_TYPES::onDispenseItem,
+                    FloatPos::newPos(pos, region.getDimensionId()),
+                    ItemClass::newItem(&const_cast<ItemStack&>(item)),
+                    Number::newNumber(slot),
+                    Number::newNumber((int)face),
+                    ContainerClass::newContainer(&container)
+                )) {
+                return;
+            }
+        }
+    }
+    IF_LISTENED_END(EVENT_TYPES::onDispenseItem);
+    origin(region, pos, face, item, container, slot, countLimit);
+}
+} // namespace dispenser
+
 namespace hopper {
 enum class HopperStatus { None, PullIn, PullOut } hopperStatus = HopperStatus::None;
 Vec3 hopperPos;
@@ -633,6 +668,7 @@ void RedstoneUpdateEvent() {
 }
 void LiquidFlowEvent() { LiquidFlowHook::hook(); }
 void CommandBlockExecuteEvent() { CommandBlockExecuteHook::hook(); }
+void DispenseItemEvent() { dispenser::DispenserEjectItemHook::hook(); }
 void HopperEvent(bool pullIn) {
     hopper::HopperAddItemHook::hook();
     if (pullIn) {