+++ src/lse/events/BlockEvents.h
@@ -11,6 +11,7 @@ void RespawnAnchorExplodeEvent();
 void PortalSpawnEvent();
 void BlockExplodedEvent();
 void RedstoneUpdateEvent();
+void DispenseItemEvent();
 void LiquidFlowEvent();
 void CommandBlockExecuteEvent();
 void HopperEvent(bool pullIn);