+++ src/tests/LSETests/EventTests.js
@@ -56,6 +56,7 @@ export const events = [
     "onFireSpread",
     "onCmdBlockExecute",
     "onContainerChange",
+    "onDispenseItem",
     "onProjectileHitBlock",
     "onRedStoneUpdate",
     "onHopperSearchItem",